requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.22.0
requests==2.31.0
readability-lxml==0.8.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai
reportlab==4.0.0
pydantic==1.10.12
